Main: Moved all functionality to Game class
